Add Cognito function to Register.jsx
