+++ .yarnrc
@@ -1,0 +1,6 @@
+# THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
+# yarn lockfile v1
+
+
+lastUpdateCheck 1580256280747
+yarn-path ".yarn\\releases\\yarn-1.21.1.js"
