--- src/components/Logout.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-// TODO Finish Logout
-
-function Logout(props) {
-    return (
-        <h1>Logout Button</h1>
-    )
-}
-
-export default Logout
